REF: Move BaseIndexers and indexer.py functions into it's own directory
