+++ pandas/api/indexers/__init__.py
@@ -2,12 +2,12 @@
 Public API for Rolling Window Indexers.
 """
 
-from pandas.core.indexers import check_array_indexer
-from pandas.core.window.indexers import (
+from pandas.core.indexers.objects import (
     BaseIndexer,
     FixedForwardWindowIndexer,
     VariableOffsetWindowIndexer,
 )
+from pandas.core.indexers.utils import check_array_indexer
 
 __all__ = [
     "check_array_indexer",
