+++ pandas/core/arrays/sparse/array.py
@@ -70,7 +70,7 @@ from pandas.core.construction import (
     extract_array,
     sanitize_array,
 )
-from pandas.core.indexers import check_array_indexer
+from pandas.core.indexers.utils import check_array_indexer
 from pandas.core.missing import interpolate_2d
 from pandas.core.nanops import check_below_min_count
 import pandas.core.ops as ops
